Basic implementation of database
With this basic implementation of database for the application we have
defined the classes we want. Further, some data have been filled into
the classes statically to check the working of the database and to learn
how to access them.

The database has been implemented based on the Room persistence library
provided by Android which abstracts database access and creation.